took table panel tests out
